fixup "base: remove Child::heap" (fix warning on 64 bit)

+++ repos/base/include/base/session_state.h
@@ -225,7 +225,7 @@ class Genode::Session_state::Factory : Noncopyable
 
 	public:
 
-		struct Batch_size { unsigned value; };
+		struct Batch_size { size_t value; };
 
 		/**
 		 * Constructor
